Added parse_lark_grammar.py and the syntax parameter/registry
