+++ tests/test_grammar.py
@@ -4,8 +4,9 @@ import os
 from unittest import TestCase, main
 
 from lark import Lark, Token, Tree, ParseError, UnexpectedInput
-from lark.load_grammar import GrammarError, GRAMMAR_ERRORS, find_grammar_errors, list_grammar_imports
+from lark.load_grammar import GrammarError, list_grammar_imports
 from lark.load_grammar import FromPackageLoader
+from lark.parse_lark_grammar import GRAMMAR_ERRORS, find_grammar_errors
 
 
 class TestGrammar(TestCase):
@@ -276,7 +277,7 @@ class TestGrammar(TestCase):
             """
 
             imports = list_grammar_imports(grammar, [os.path.dirname(__file__)])
-            self.assertEqual({os.path.split(i)[-1] for i in imports}, {'test_templates_import.lark', 'templates.lark'})
+            self.assertEqual({'test_templates_import.lark', 'templates.lark'}, {os.path.split(i)[-1] for i in imports})
 
             imports = list_grammar_imports('%import common.WS', [])
             assert len(imports) == 1 and imports[0].pkg_name == 'lark'
